% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPseudoBulk.avg.R
\name{flt.cluster}
\alias{flt.cluster}
\title{Title}
\usage{
flt.cluster(x, th.exp.pct, th.drp.pct, th.drp.cell)
}
\arguments{
\item{x}{data, int: vector of expression values}

\item{th.exp.pct}{threshold, int: quantile cap of expression values within a cluster to remove spikes in expression (default = 90)}

\item{th.drp.pct}{threshold, int: minimum percent of dropout cells within a cell cluster below which expression values are set to 0 (default = 10)}

\item{th.drp.cell}{threshold, int: minimum number of dropout cells within a cell cluster below which expression values are set to 0 (default = 3)}
}
\value{

}
\description{
Title
}
