% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPseudoBulk.avg.R
\name{scPseudoBulk.avg}
\alias{scPseudoBulk.avg}
\title{scPseudoBulk.avg}
\usage{
scPseudoBulk.avg(
  norm = NULL,
  clusters = NULL,
  th.exp.pct = 90,
  th.drp.pct = 10,
  th.drp.cell = 3
)
}
\arguments{
\item{norm}{data, matrix/dataframe: Normalized count matrix (rows = genes, columns = cells)}

\item{clusters}{data, factors: vector of clusters for each cell}

\item{th.exp.pct}{threshold, int: quantile cap of expression values within a cluster to remove spikes in expression (default = 90)}

\item{th.drp.pct}{threshold, int: minimum percent of dropout cells within a cell cluster below which expression values are set to 0 (default = 10)}

\item{th.drp.cell}{threshold, int: minimum number of dropout cells within a cell cluster below which expression values are set to 0 (default = 3)}
}
\value{

}
\description{
scPseudoBulk.avg
}
